refactor: brush geo toolbox.
